requirements.txt
python-dotenv
selenium
pymongo
dnspython
